Routing K-means/app/requirements.txt
aniso8601==8.0.0
certifi==2020.6.20
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.2
python-dateutil==2.8.1
python-logstash==0.4.6
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
wincertstore==0.2
geopy
